requirements.txt
opencv-python
mediapipe
numpy
ultralytics
Pillow
streamlit
matplotlib
